Refactor: Separate utility functions and improve code organization
