+++ src/utils.c
@@ -12,7 +12,7 @@
 
 #include "../pipex.h"
 
-static void	free_paths(char **paths)
+void	free_paths(char **paths)
 {
 	int	i;
 
@@ -22,7 +22,7 @@ static void	free_paths(char **paths)
 	free(paths);
 }
 
-static char	*join_path(const char *dir, const char *cmd)
+char	*join_path(const char *dir, const char *cmd)
 {
 	char	*part_path;
 	char	*full_path;
@@ -66,7 +66,7 @@ void	ft_error(const char *msg)
 	perror(msg);
 }
 
-static void	free_cmd(char **cmd)
+void	free_cmd(char **cmd)
 {
 	int	i;
 
@@ -104,31 +104,4 @@ void	execute(char *argv, char **envp)
 		ft_error("execve");
 		exit(126);
 	}
-}
-
-int	get_next_line(char **line)
-{
-	char	*buffer;
-	int		i;
-	int		r;
-	char	c;
-
-	i = 0;
-	r = 0;
-	buffer = (char *)malloc(10000);
-	if (!buffer)
-		return (-1);
-	r = read(0, &c, 1);
-	while (r && c != '\n' && c != '\0')
-	{
-		if (c != '\n' && c != '\0')
-			buffer[i] = c;
-		i++;
-		r = read(0, &c, 1);
-	}
-	buffer[i] = '\n';
-	buffer[++i] = '\0';
-	*line = buffer;
-	free(buffer);
-	return (r);
 }
